+++ changelog.md
@@ -1,5 +1,6 @@
 # 1.1.3
 - Added brightness settings to the rainbow effect
+- Added particle rainbow effect (toggleable)
 
 # 1.1.2
 - Fixed the mod not working with the new update